rename HelloWorld component to Homepage
